development_requirements.txt
# Only needed for syntax highlighting
fake-bpy-module
